feat: Add support for browser-based logins
For sites that only support logging in using a third-party login provider, allow logging in via a browser.
This new flow will open a custom tab where a user can log in using the supported mechanisms of the site,
then have the client id redirect back to to the app using a callback URL.
The auth code returned by the callback URL can then be used instead of a username and password to log in.